Clean up, localize and add icons
* Change component name to `LocationInstructionsAlert`
* Add helper icons and browser icons for read-friendly text
* Dynamically update button link context based on UA device and version
* Translate localized text
