+++ webapp/public/locale/es/strings.json
@@ -81,7 +81,7 @@
     "checkout_fatal_error": "Lo sentimos, algo salió mal al intentar comprobarlo. Por favor, regrese y vuelva a intentarlo.",
     "checkout_no_items": "No hay artículos en su carrito, o ninguno de sus artículos está disponible. Por favor, regrese y vuelva a intentarlo.",
     "conflicting_bank_account": "Tienes una cuenta bancaria con ese número de cuenta y clave bancaria ya. Intenta de nuevo, o entre los datos de otra cuenta.",
-    "denied_geolocation": "Debe activar la geolocalización en tu depositivo para usar la app de movilidad de Suma.",
+    "denied_geolocation": "Debe permitir servicios de localización en tu aparato antes de usar los Servicios de Movilidad de suma.",
     "disambiguation_required": "$t(strings:errors:unhandled_error)",
     "eligibility_violation": "$t(strings:errors:unhandled_error)",
     "enable_geolocation_instructions": "Cómo activar la geolocalización",
@@ -239,6 +239,15 @@
     "refund": "{{memo}}",
     "unknown": "Tarifa de Suma no clasificada"
   },
+  "location_instructions_links": {
+    "chrome": "https://support.google.com/chrome/answer/142065?hl={{lang}}&co=GENIE.Platform%3D{{platform}}&oco=1",
+    "edge": "https://support.microsoft.com/{{lang}}-us/windows/windows-location-service-and-privacy-3a8eee0a-5b0b-dc07-eede-2a5ca1c49088",
+    "firefox": "https://support.mozilla.org/{{lang}}-US/kb/does-firefox-share-my-location-websites",
+    "ios": "https://support.apple.com/{{lang}}-us/HT207092",
+    "label": "Instrucciones de {{device}}",
+    "label_with_version": "Instrucciones de {{device}} {{version}}",
+    "safari": "https://support.apple.com/guide/mac-help/allow-apps-to-detect-the-location-of-your-mac-mh35873/mac#:~:text=use%20Location%20Services-,On%20your%20Mac%2C%20choose%20Apple%20menu%20%3E%20System%20Settings%2C%20click,may%20need%20to%20scroll%20down.\\)&text=Specify%20which%20apps%20and%20system%20services%20can%20use%20Location%20Services"
+  },
   "mobility": {
     "do_not_park_intro": "No se estacione en esta zona. Podemos tomar medidas por violaciones repetidas.",
     "do_not_park_title": "Por favor no se estacione aquí.",
@@ -249,9 +258,10 @@
     "locate_me": "Localizarme",
     "mobility_lime_access_summer_2023_rate": "Gratis durante 30 minutos, luego {{unitCents, sumaCurrency}} por minuto",
     "mobility_start_and_per_minute": "{{surchargeCents, sumaCurrency}} para iniciar, {{unitCents, sumaCurrency}} por minuto",
+    "reload_after_instructions": "**Refrescar** esta página una vez que haya terminado de seguir las instrucciones.",
     "reserve_scooter": "Reservar una patineta eléctrica",
     "setup_private_account_with_vendor": "Para comenzar, configuraremos una cuenta privada para usted con {{vendorName}}.",
-    "title": "Servicios de movilidad",
+    "title": "Servicios de Movilidad",
     "trip_ended": "Tu viaje con {{vendor}} ah terminado.\n\nUsamos {{totalCost, sumaCurrency}} de su saldo de Suma.\n\nSalvaste {{discountAmount, sumaCurrency}}.",
     "trip_started_at": "Viaje iniciado a {{at}}.",
     "zoom_in": "Acercar",
@@ -351,14 +361,6 @@
     "terms_of_sale_holiday_2022": "Condiciones de Venta (Días Festivos de 2022)",
     "terms_of_sale_summer_2023": "Condiciones de Venta (Verano 2023)",
     "terms_of_use": "Términos de Uso"
-  },
-  "location_instructions_links": {
-    "label": "Instrucciones de {{device}}",
-    "firefox": "https://support.mozilla.org/{{lang}}-US/kb/does-firefox-share-my-location-websites",
-    "ios": "https://support.apple.com/{{lang}}-us/HT207092",
-    "safari": "https://support.apple.com/guide/mac-help/allow-apps-to-detect-the-location-of-your-mac-mh35873/mac#:~:text=use%20Location%20Services-,On%20your%20Mac%2C%20choose%20Apple%20menu%20%3E%20System%20Settings%2C%20click,may%20need%20to%20scroll%20down.\\)&text=Turn%20Location%20Services%20on%20or,the%20list%20on%20the%20right.",
-    "chrome": "https://support.google.com/chrome/answer/142065?hl={{lang}}&co=GENIE.Platform%3D{{platform}}&oco=1",
-    "edge": "https://support.microsoft.com/{{lang}}-us/windows/windows-location-service-and-privacy-3a8eee0a-5b0b-dc07-eede-2a5ca1c49088"
   },
   "utilities": {
     "intro": "Suma quiere facilitarle el ahorro de dinero en sus facturas de servicios públicos y hacer que nuestras comunidades sean más resilientes frente al clima extremo y el cambio climático.\n\nRegístrese en nuestra lista de espera gratuita para ayudarnos a construir una comunidad de miembros y demostrar que todos merecen energía limpia y asequible.",
