+++ webapp/src/components/mobilitymap/Map.js
@@ -8,7 +8,7 @@ import { useGlobalViewState } from "../../state/useGlobalViewState";
 import { useUser } from "../../state/useUser";
 import FormError from "../FormError";
 import CardOverlay from "./CardOverlay";
-import GeolocationInstructionsModal from "./GeolocationInstructionsModal";
+import LocationInstructionsAlert from "./LocationInstructionsAlert";
 import ReservationCard from "./ReservationCard";
 import TripCard from "./TripCard";
 import React from "react";
@@ -61,7 +61,7 @@ const Map = () => {
     []
   );
   const handleLocationError = React.useCallback(
-    () => setError(<GeolocationInstructionsModal />),
+    () => setError(<LocationInstructionsAlert />),
     [setError]
   );
